refactor : component import 리팩터
