+++ src/component/Personal/Add/index.tsx
@@ -1,0 +1,5 @@
+import PhotoUpload from './PhotoUpload';
+import PreviewPhoto from './PreviewPhoto';
+import Writer from './Writer';
+
+export { PhotoUpload, PreviewPhoto, Writer };
