+++ src/component/Personal/index.tsx
@@ -1,0 +1,17 @@
+import PhotoPreview from '../Common/PhotoPreview';
+import BottomAnimationImage from './BottomAnimationImage';
+import NoneContent from './NoneContent';
+import OwnerButtonLayer from './OwnerButtonLayer';
+import SnowAnimation from './SnowAnimation';
+import Summary from './Summary';
+import WriterButtonLayer from './WriterButtonLayer';
+
+export {
+  BottomAnimationImage,
+  NoneContent,
+  OwnerButtonLayer,
+  SnowAnimation,
+  Summary,
+  WriterButtonLayer,
+  PhotoPreview,
+};
